refactor(interface): extract useShieldFlow — ShieldModal becomes a dumb renderer (no behavior change)
